pyproject.toml
[project]
name = "pokedex"
version = "0.1.0"
description = "A simple cli tool for printing ascii art of any pokemon."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.1.1",
]

[project.scripts]
pokedex = "pokedex.cli:main"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
